Move item listing models to common location for reuse with search
